+++ src/Espresso.java
@@ -1,8 +1,8 @@
 public class Espresso implements Coffee{
 
     @Override
-    public int getCost() {
-        return 3;
+    public double getCost() {
+        return 3.0;
     }
 
     @Override
